# apps/web/src/lib/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "louped",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/playground/load": {
      "post": {
        "summary": "Load Model",
        "description": "Replace the model (none unloads it, freeing the GPU for jobs); a bad one is a 400 and\nleaves no model loaded.",
        "operationId": "load_model_api_playground_load_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaygroundInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/playground/generate": {
      "post": {
        "summary": "Generate",
        "operationId": "generate_api_playground_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/playground/inspect": {
      "post": {
        "summary": "Inspect",
        "operationId": "inspect_api_playground_inspect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InspectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/playground/patch": {
      "post": {
        "summary": "Patch",
        "operationId": "patch_api_playground_patch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/playground/dose": {
      "post": {
        "summary": "Dose",
        "operationId": "dose_api_playground_dose_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DoseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/playground/speed": {
      "post": {
        "summary": "Speed",
        "operationId": "speed_api_playground_speed_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeedRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DoseRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Prompt"
          },
          "vector": {
            "type": "string",
            "title": "Vector"
          },
          "layer": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Layer"
          },
          "alphas": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "maxItems": 41,
            "minItems": 2,
            "title": "Alphas"
          },
          "answer": {
            "type": "string",
            "minLength": 1,
            "title": "Answer"
          },
          "foil": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Foil"
          },
          "adapters": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Adapters",
            "default": []
          },
          "chat": {
            "type": "boolean",
            "title": "Chat",
            "default": true
          }
        },
        "type": "object",
        "required": ["prompt", "vector", "alphas", "answer"],
        "title": "DoseRequest",
        "description": "A saved direction swept over strengths, read at the next token."
      },
...
      "GenerateRequest": {
        "properties": {
          "interventions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Interventions",
            "default": []
          },
          "adapters": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Adapters",
            "description": "Adapters of the bank live for this request.",
            "default": []
          },
          "max_new_tokens": {
            "type": "integer",
            "maximum": 512.0,
            "minimum": 1.0,
            "title": "Max New Tokens",
            "description": "Tokens to generate; a diffusion model's reply length.",
            "default": 64
          },
          "steps": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 512.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Steps",
            "description": "Denoising steps; the length if unset."
          },
          "prompt": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Prompt"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/Turn"
            },
            "type": "array",
            "maxItems": 16,
            "title": "History",
            "description": "Earlier turns, before prompt: a follow-up such as pushback.",
            "default": []
          }
        },
        "type": "object",
        "required": ["prompt"],
        "title": "GenerateRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InspectRequest": {
        "properties": {
          "interventions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Interventions",
            "default": []
          },
          "adapters": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Adapters",
            "description": "Adapters of the bank live for this request.",
            "default": []
          },
          "max_new_tokens": {
            "type": "integer",
            "maximum": 512.0,
            "minimum": 1.0,
            "title": "Max New Tokens",
            "description": "Tokens to generate; a diffusion model's reply length.",
            "default": 64
          },
          "steps": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 512.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Steps",
            "description": "Denoising steps; the length if unset."
          },
          "prompt": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Prompt"
          },
          "vectors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Vectors",
            "description": "Saved directions to read, each at its own layer.",
            "default": []
          },
          "chat": {
            "type": "boolean",
            "title": "Chat",
            "description": "Wrap the prompt in the model's chat template.",
            "default": true
          }
        },
        "type": "object",
        "required": ["prompt"],
        "title": "InspectRequest"
      },
      "InspectResponse": {
        "properties": {
          "views": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/HeatmapView"
                },
                {
                  "$ref": "#/components/schemas/LineView"
                },
                {
                  "$ref": "#/components/schemas/ScatterView"
                },
                {
                  "$ref": "#/components/schemas/TableView"
                },
                {
                  "$ref": "#/components/schemas/TokensView"
                }
              ],
              "discriminator": {
                "propertyName": "kind",
                "mapping": {
                  "heatmap": "#/components/schemas/HeatmapView",
                  "line": "#/components/schemas/LineView",
                  "scatter": "#/components/schemas/ScatterView",
                  "table": "#/components/schemas/TableView",
                  "tokens": "#/components/schemas/TokensView"
                }
              }
            },
            "type": "array",
            "title": "Views"
          }
        },
        "type": "object",
        "required": ["views"],
        "title": "InspectResponse"
      },
...
      "LoadRequest": {
        "properties": {
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "bank": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Bank",
            "default": []
          },
          "diffusion": {
            "type": "boolean",
            "title": "Diffusion",
            "default": false
          },
          "attn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attn"
          },
          "remote_code": {
            "type": "boolean",
            "title": "Remote Code",
            "default": false
          }
        },
        "type": "object",
        "title": "LoadRequest",
        "description": "A model for the Playground: a Hub id, a path or a name under <home>/models; adapters to load\nbeside it from <home>/adapters; whether it is a masked diffusion model; its attention kernel\n(eager, sdpa, flash_attention_2, flex_attention, a registered name, or file.py:function)."
      },
...
      "PatchRequest": {
        "properties": {
          "clean": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Clean"
          },
          "corrupt": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Corrupt"
          },
          "answer": {
            "type": "string",
            "minLength": 1,
            "title": "Answer",
            "description": "The clean prompt's next token."
          },
          "foil": {
            "type": "string",
            "minLength": 1,
            "title": "Foil",
            "description": "The corrupt prompt's next token."
          },
          "method": {
            "type": "string",
            "enum": ["attribution", "residual", "heads"],
            "title": "Method",
            "default": "attribution"
          },
          "adapters": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Adapters",
            "default": []
          },
          "chat": {
            "type": "boolean",
            "title": "Chat",
            "default": true
          }
        },
        "type": "object",
        "required": ["clean", "corrupt", "answer", "foil"],
        "title": "PatchRequest",
        "description": "Which layer, position or head carries the difference between two prompts of one length."
      },
      "PlaygroundInfo": {
        "properties": {
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "layers": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Layers"
          },
          "heads": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heads"
          },
          "bank": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Bank",
            "default": []
          },
          "diffusion": {
            "type": "boolean",
            "title": "Diffusion",
            "default": false
          },
          "switchable": {
            "type": "boolean",
            "title": "Switchable",
            "default": false
          }
        },
        "type": "object",
        "required": ["model", "layers"],
        "title": "PlaygroundInfo"
      },
...
      "Turn": {
        "properties": {
          "role": {
            "type": "string",
            "enum": ["user", "assistant"],
            "title": "Role"
          },
          "content": {
            "type": "string",
            "maxLength": 8000,
            "title": "Content"
          }
        },
        "type": "object",
        "required": ["role", "content"],
        "title": "Turn"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}